simplify logger
