+++ packages/logger/src/interfaces/InfoData.ts
@@ -1,0 +1,10 @@
+import type { PackageTaskInfo } from "./PackageTaskInfo";
+
+/**
+ * LogStructuredData for the `info` command
+ */
+export interface InfoData {
+  command?: string[];
+  scope?: string[];
+  packageTasks?: PackageTaskInfo[];
+}
