+++ packages/logger/src/interfaces/LogEntry.ts
@@ -1,0 +1,10 @@
+import type { LogLevel } from "./LogLevel";
+import type { LogStructuredData } from "./LogStructuredData";
+
+
+export interface LogEntry {
+  timestamp: number;
+  level: LogLevel;
+  msg: string;
+  data?: LogStructuredData;
+}
