+++ packages/logger/src/interfaces/LogLevel.ts
@@ -1,0 +1,9 @@
+export const LogLevel = {
+  error: 10,
+  warn: 20,
+  info: 30,
+  verbose: 40,
+  silly: 50,
+} as const;
+
+export type LogLevel = typeof LogLevel[keyof typeof LogLevel];
