+++ packages/logger/src/interfaces/LogStructuredData.ts
@@ -1,0 +1,5 @@
+import { InfoData } from "./InfoData";
+import { TaskData } from "./TaskData";
+
+
+export type LogStructuredData = TaskData | InfoData;
