+++ packages/logger/src/interfaces/TaskData.ts
@@ -1,0 +1,7 @@
+export interface TaskData {
+  status?: "pending" | "started" | "completed" | "failed" | "skipped";
+  package?: string;
+  task?: string;
+  duration?: string;
+  hash?: string | null;
+}
